+++ src_new/triggers/levi_ataxia/for_levi/leviticus/highlighting/034_Breath_Echoes.lua
@@ -1,0 +1,57 @@
+--[[mudlet
+type: trigger
+name: Breath Echoes
+hierarchy:
+- Levi_Ataxia
+- For Levi
+- leviticus
+- LeviAtax
+- Leviticus
+- Highlighting
+attributes:
+  isActive: 'yes'
+  isFolder: 'no'
+  isTempTrigger: 'no'
+  isMultiline: 'no'
+  isPerlSlashGOption: 'no'
+  isColorizerTrigger: 'no'
+  isFilterTrigger: 'no'
+  isSoundTrigger: 'no'
+  isColorTrigger: 'no'
+  isColorTriggerFg: 'no'
+  isColorTriggerBg: 'no'
+triggerType: 0
+conditonLineDelta: 0
+mStayOpen: 0
+mCommand: ''
+packageName: ''
+mFgColor: '#ff0000'
+mBgColor: '#ffff00'
+mSoundFile: ''
+colorTriggerFgColor: '#000000'
+colorTriggerBgColor: '#000000'
+patterns:
+- pattern: ^An echo of psionic breath assails (.+), forcing \w+ to stumble clumsily around\.$
+  type: 1
+- pattern: ^An echo of flaming breath strikes (.+), scorching \w+ and inhibiting \w+ natural healing\.$
+  type: 1
+- pattern: ^An echo of frigid breath chills (.+), slowing \w+ movements to a crawl\.$
+  type: 1
+]]--
+
+-- Dragon breath-echo procs (live capture 2026-07-28, source unconfirmed --
+-- boon/affix?): the psionic echo makes the denizen CLUMSY (stumbling), the flaming
+-- echo INHIBITS its natural healing, the frigid echo gives it AEON (slowed to a
+-- crawl). Distinct colours so the effects read at a glance: psionic = orchid
+-- (control family), flaming = tomato (burn family), frigid = cyan (ice family).
+selectString(line,1)
+setBold(true)
+if line:find("psionic") then
+	fg("orchid")
+elseif line:find("frigid") then
+	fg("cyan")
+else
+	fg("tomato")
+end
+deselect()
+resetFormat()